[Jacob] Refactored and documented backend
