+++ backend/README.md
@@ -1,0 +1,87 @@
+# Backend
+
+## Setup 
+
+### Docker
+
+Run `docker-compose up -d api` from the project root
+
+### Manual
+
+```shell
+pip install -r requirements.txt
+uvicorn app.main:app --reload --port 8000
+```
+
+
+
+## Usage
+
+To examine and try out the different endpoints, go to [`localhost:8000/docs`](localhost:8000/docs) for the OpenAPI documentation and [`localhost:8000/redocs`](localhost:8000/redocs) for the ReDocs, both automatically generated by `fastapi`.
+
+
+### Python
+
+```python
+import requests
+res = requests.post('localhost:8000/subgraph', data={'node_ids': [431, 1198, 828, 59, 1206], 'k': 2})
+data = res.json()
+```
+
+### CuRL
+
+```shell
+curl -X 'POST' \
+  'http://localhost:8000/subgraph' \
+  -H 'accept: application/json' \
+  -H 'Content-Type: application/json' \
+  -d '{
+  "node_ids": [431, 1198, 828, 59, 1206],
+  "k": 2
+}'
+```
+
+### Example Data
+
+```javascript
+{
+    "numNodes": 209,
+    "nodes": [
+        {
+            "nhs_no": "632-68-0917",
+            "surname": "Rice",
+            "name": "Anne",
+            "label": "Person",
+            "id": "431"
+        },
+        {
+            "nhs_no": "351-83-4608",
+            "surname": "Montgomery",
+            "name": "Andrea",
+            "label": "Person",
+            "id": "1198"
+        },
+        {
+            "nhs_no": "271-78-8919",
+            "surname": "Willis",
+            "name": "Arthur",
+            "label": "Person",
+            "id": "828"
+        } 
+        // ...
+    ],
+    "edges": [
+        {
+            "source": "431",
+            "label": "CURRENT_ADDRESS",
+            "target": "432"
+        },
+        {
+            "rel_type": "PARENT",
+            "source": "1198",
+            "label": "FAMILY_REL",
+            "target": "1067"
+        } // ...
+    ]
+}
+```
