+++ backend/app/models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Optional, List
+
+
+class Subgraph(BaseModel):
+    node_ids: List[int] = []
+    k: Optional[int] = 1
