+++ docker-compose.yml
@@ -32,7 +32,7 @@ services:
     environment:
       - NEO4J_URI=neo4j://neo4j:7687
       - NEO4J_AUTH=${NEO4J_AUTH}
-    command: uvicorn app:app --host 0.0.0.0 --reload
+    command: uvicorn app.main:app --host 0.0.0.0 --reload
     depends_on:
       - neo4j
