+++ web-react/package.json
@@ -9,6 +9,8 @@
     "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.60",
+    "@mui/material": "^5.2.1",
+    "@mui/x-data-grid": "^5.0.1",
     "classnames": "^2.2.6",
     "graphql": "^14.5.8",
     "moment": "^2.29.1",
